feat(frontend): RSC split, hydration-aware loading, and i18n for UserPermissionsManager
Splits the component into a server entry point and a "use client" island,
tracks zustand persist hydration to avoid a flash of empty state, and adds
translated strings for the permissions manager UI.
